Squid_Game & Video demo presentation
Initially, the program rolls a magic die from 1 to 12 and informs the user with a message that they have 4 attempts to guess it.[eos] Next, the user makes their first attempt. If they guess the number correctly, a message is displayed saying “You Win!”. Otherwise, the user is informed that they guessed incorrectly and must choose their next number. If they run out of attempts without finding the number, they are informed that they lost. If the user provides invalid input (e.g., a string) when asked to choose a number, the attempt is void, their remaining attempts are not reduced, and they are asked to enter a number again.